+++ tools/gulp/tasks/clean.ts
@@ -1,6 +1,6 @@
 import {task} from 'gulp';
 import {DIST_ROOT} from '../constants';
-import {cleanTask} from '../task_helpers';
+import {cleanTask} from '../util/task_helpers';
 
 
 task('clean', cleanTask(DIST_ROOT));
